added user usdc early withdraw

+++ src/interfaces/IHippodrome.sol
@@ -21,7 +21,7 @@ interface IHippodrome is IHippodromeTypes{
     function fundCampaign(uint128 campaignId, uint amount) external virtual;
     function withdrawFunds(uint128 campaign, uint amount) external virtual;
     function claimRewards(uint128 campaignID) external virtual;
-    function resolveCampaign(uint campaignID) external virtual;
+    function resolveCampaign(uint128 campaignID) external virtual;
 
     function getAvailableUserRewards(address user, uint128 campaignID) external view virtual returns (uint);
     function calculateContributionPercentage(uint128 campaignID, address user) external view virtual returns (uint);
